requirements.txt
SmartApi==1.5.5
pyotp==2.9.0
logzero==1.7.0
python-dotenv==1.0.1
